% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfo_om.R
\name{transfo_om}
\alias{transfo_om}
\title{Transformer un territoire de France d'Outre-Mer pour le placer près de la France métropolitaine.}
\usage{
transfo_om(shape_origine, var_departement, type_transfo)
}
\arguments{
\item{shape_origine}{Table géographique (format sf dataframe) en entrée contenant les entités du territoire de France d'Outre-Mer.}

\item{var_departement}{Nom du champ de la table géographique en entrée contenant le code du département de France d'Outre-Mer.}

\item{type_transfo}{Version de la transformation à appliquer.}
}
\value{
Renvoie une table géographique (format sf dataframe) contenant les entités du territoire de France d'Outre-Mer replacés.
}
\description{
Modifie le champ géometrie du sf dataframe en entrée.
}
